Major Refactoring
Moved the different layers out in their seperate programs.
